Fixed Sim distance calc, ShooterCalculatorTest
Also moved DriverStation.getInstance() into BitBucketsSubsystem constructor for unit tests.
